Adjusted table and transaction js. Gave the table a clearer parent component, removed fake data. Removed some old comments and such. Converted date format.
