+++ northark-frontend/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import Layout from './components/Layout';
-import Table from './components/Table';
+import Table from './components/Transaction';
 
 
 ReactDOM.render(
